# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Workforce API",
    "version": "1.0.0"
  },
  "paths": {
    "/flows": {
      "get": {
        "summary": "Get all flows",
        "responses": {
          "200": {
            "description": "A list of flows",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/flow"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new flow",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/flow"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The created flow",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/flow"
                }
              }
            }
          }
        }
      }
    },
    "/flows/{id}": {
      "get": {
        "summary": "Get a flow by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The flow",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/flow"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a flow by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/flow"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated flow",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/flow"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a flow by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The deleted flow",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/flow"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "flow": {
        "allOf": [
          {
            "$ref": "#/components/schemas/object"
          },
          {
            "type": "object",
            "properties": {
              "status": {
                "type": "string",
                "enum": [
                  "active",
                  "inactive"
                ]
              },
              "channels": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/channel"
                }
              },
              "documentation": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/documentation"
                }
              },
              "resources": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/resource"
                }
              },
              "tasks": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/task"
                }
              },
              "tools": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/tool"
                }
              },
              "trackers": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/tracker"
                }
              }
            },
            "required": [
              "status"
            ]
          }
        ]
      },
...
      "channel": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "The type of the channel",
            "enum": [
              "discord-channel",
              "native-channel",
              "slack-channel"
            ]
          },
          "credential": {
            "type": "string"
          },
          "variables": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/discord_channel_variables"
              },
              {
                "$ref": "#/components/schemas/native_channel_variables"
              },
              {
                "$ref": "#/components/schemas/slack_channel_variables"
              }
            ]
          }
        },
        "required": [
          "type",
          "credential",
          "variables"
        ]
      },
...
      "documentation": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "The type of the documentation",
            "enum": [
              "default-documentation"
            ]
          },
          "repository": {
            "type": "string",
            "description": "The name of the document repository to use."
          },
          "documents": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The list of documents from the repository to use. Null or Empty means all"
          },
          "variables": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/default_documentation_variables"
              }
            ]
          }
        },
        "required": [
          "type",
          "repository",
          "variables"
        ]
      },
...
      "resource": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "The type of the resource",
            "enum": [
              "api-resource",
              "github-pull-request-resource",
              "github-repo-resource",
              "raw-text-resource"
            ]
          },
          "credential": {
            "type": "string"
          },
          "example": {
            "type": "string",
            "description": "An example value of an output for the resource."
          },
          "variables": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/api_resource_variables"
              },
              {
                "$ref": "#/components/schemas/github_pull_request_resource_variables"
              },
              {
                "$ref": "#/components/schemas/github_repo_resource_variables"
              },
              {
                "$ref": "#/components/schemas/raw_text_resource_variables"
              }
            ]
          }
        },
        "required": [
          "type",
          "credential",
          "variables"
        ]
      },
      "task": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "The type of the task",
            "enum": [
              "simple-task",
              "structured-task"
            ]
          },
          "requiredSkills": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The list of skills needed for this task."
          },
          "defaultChannel": {
            "type": "string",
            "description": "The default channel to use for communication."
          },
          "tracker": {
            "type": "string",
            "description": "The Name of the tracker to use for execution."
          },
          "documentation": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The list of documentation names to use."
          },
          "tools": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/tool_reference"
            },
            "description": "The list of tools needed for this task."
          },
          "triggers": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The list of channels or resources to trigger the task."
          },
          "inputs": {
            "type": "object",
            "description": "The map of inputs for this task. Keys can be used in task templates using handlebars syntax. I.E., and input of message may map to a channel name, and then in the task template, you can use {{message}}",
            "additionalProperties": {
              "oneOf": [
                {
                  "type": "string"
                },
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              ]
            }
          },
          "outputs": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The list of outputs for this task."
          },
          "subtasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/subtask"
            },
            "description": "The list of subtasks for this task."
          },
          "costLimit": {
            "type": "number",
            "description": "The cost limit for each execution of this task."
          },
          "variables": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/simple_task_variables"
              },
              {
                "$ref": "#/components/schemas/structured_task_variables"
              }
            ]
          }
        },
        "required": [
          "type",
          "variables"
        ]
      },
      "tool": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "The type of the tool",
            "enum": [
              "coding-tool",
              "excalidraw-tool",
              "github-board-ticket-tool",
              "google-drive-tool",
              "google-slides-tool",
              "message-channel-tool",
              "openapi-channel-tool",
              "openapi-tool",
              "template-tool",
              "trello-ticket-tool",
              "web-service-tool"
            ]
          },
          "credential": {
            "type": "string"
          },
          "channel": {
            "type": "string",
            "description": "The channel for the tool to use for dynamic user interaction"
          },
          "variables": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/coding_tool_variables"
              },
              {
                "$ref": "#/components/schemas/excalidraw_tool_variables"
              },
              {
                "$ref": "#/components/schemas/github_board_ticket_tool_variables"
              },
              {
                "$ref": "#/components/schemas/google_drive_tool_variables"
              },
              {
                "$ref": "#/components/schemas/google_slides_tool_variables"
              },
              {
                "$ref": "#/components/schemas/message_channel_tool_variables"
              },
              {
                "$ref": "#/components/schemas/openapi_channel_tool_variables"
              },
              {
                "$ref": "#/components/schemas/openapi_tool_variables"
              },
              {
                "$ref": "#/components/schemas/template_tool_variables"
              },
              {
                "$ref": "#/components/schemas/trello_ticket_tool_variables"
              },
              {
                "$ref": "#/components/schemas/web_service_tool_variables"
              }
            ]
          }
        },
        "required": [
          "type",
          "credential",
          "variables"
        ]
      },
      "tracker": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "The type of the tracker",
            "enum": [
              "github-board-tracker",
              "trello-tracker"
            ]
          },
          "credential": {
            "type": "string"
          },
          "webhooksEnabled": {
            "type": "boolean",
            "description": "Whether or not webhooks are enabled for this tracker."
          },
          "pollingInterval": {
            "type": "number",
            "description": "The interval in seconds to poll for updates if webhooks are not enabled"
          },
          "variables": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/github_board_tracker_variables"
              },
              {
                "$ref": "#/components/schemas/trello_tracker_variables"
              }
            ]
          }
        },
        "required": [
          "type",
          "credential",
          "variables"
        ]
      },
...
      "object": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "The unique identifier for the object"
          },
          "orgId": {
            "type": "string",
            "format": "uuid",
            "description": "The unique identifier for the organization the object belongs to"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "description"
        ]
      },
      "tool_reference": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "output": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "subtask": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "async": {
            "type": "boolean"
          }
        },
        "required": [
          "name"
        ]
      },
      "discord_channel_variables": {
        "type": "object",
        "properties": {
          "channel_id": {
            "type": "string",
            "description": "The Discord channel to send messages to",
            "x-workforce-sensitive": false
          }
        },
        "required": [
          "channel_id"
        ]
      },
      "native_channel_variables": {
        "type": "object",
        "properties": {
          "channel_name": {
            "type": "string",
            "description": "A unique name for the channel",
            "x-workforce-sensitive": false
          },
          "anonymous_access": {
            "type": "boolean",
            "description": "Allow anonymous access to the channel",
            "x-workforce-sensitive": false
          },
          "voice_enabled": {
            "type": "boolean",
            "description": "Enable voice input for the channel",
            "x-workforce-sensitive": false,
            "x-workforce-advanced": true
          },
          "oauth2_issuer_uri": {
            "type": "string",
            "description": "Tokens from this issuer will be accepted",
            "x-workforce-sensitive": false
          },
          "oauth2_audience": {
            "type": "string",
            "description": "The audience for the token",
            "x-workforce-sensitive": false
          },
          "oauth2_claims": {
            "type": "string",
            "description": "A JSON object containing the claims to be verified",
            "x-workforce-multiline": true,
            "x-workforce-sensitive": false
          }
        },
        "required": [
          "channel_name"
        ]
      },
      "slack_channel_variables": {
        "type": "object",
        "properties": {
          "channel_id": {
            "type": "string",
            "description": "The Slack channel to send messages to",
            "x-workforce-sensitive": false
          }
        },
        "required": [
          "channel_id"
        ]
      },
...
      "default_documentation_variables": {
        "type": "object",
        "properties": {
          "desiredTokens": {
            "type": "number",
            "description": "The desired number of tokens to generate.",
            "default": 512
          },
          "maxTokens": {
            "type": "number",
            "description": "The maximum number of tokens to generate.",
            "default": 1024
          },
          "retrievalScope": {
            "type": "string",
            "description": "The scope of the retrieval.",
            "default": "section",
            "enum": [
              "repository",
              "document",
              "section",
              "chunk"
            ]
          },
          "tokenFillStrategy": {
            "type": "string",
            "description": "The strategy for filling tokens.",
            "default": "default",
            "enum": [
              "default",
              "chunk_first",
              "fill_section",
              "fill_document"
            ]
          }
        },
        "required": []
      },
...
      "api_resource_variables": {
        "type": "object",
        "properties": {
          "schema_url": {
            "type": "string",
            "description": "A URL to a valid OpenAPI schema."
          },
          "raw_schema": {
            "type": "string",
            "description": "A valid OpenAPI schema.",
            "x-workforce-multiline": true
          },
          "fetch_path": {
            "type": "string",
            "description": "The path to use for fetching data from the API."
          },
          "fetch_method": {
            "type": "string",
            "description": "The HTTP method to use for fetching data from the API.",
            "enum": [
              "GET",
              "POST",
              "PUT",
              "PATCH",
              "DELETE"
            ]
          },
          "webhook_path": {
            "type": "string",
            "description": "The path that defines the webhook endpoint."
          },
          "create_path": {
            "type": "string",
            "description": "The path to use for creating new objects in the API."
          },
          "create_method": {
            "type": "string",
            "description": "The HTTP method to use for creating new objects in the API.",
            "enum": [
              "GET",
              "POST",
              "PUT",
              "PATCH",
              "DELETE"
            ]
          }
        },
        "required": []
      },
      "github_pull_request_resource_variables": {
        "type": "object",
        "properties": {
          "repo": {
            "type": "string",
            "description": "The name(slug) of the Github repository to use."
          },
          "owner": {
            "type": "string",
            "description": "The name(slug) of owner of the Github repository."
          }
        },
        "required": [
          "repo",
          "owner"
        ]
      },
      "github_repo_resource_variables": {
        "type": "object",
        "properties": {
          "repo": {
            "type": "string",
            "description": "The name(slug) of the Github repository to use."
          },
          "owner": {
            "type": "string",
            "description": "The name(slug) of owner of the Github repository."
          },
          "branch": {
            "type": "string",
            "description": "The branch name.",
            "default": "main"
          },
          "path_template": {
            "type": "string",
            "description": "The path template to use to generate the path to the file. This can also just be the name of a particular or directory. Defaults to {{filename}}.",
            "default": "{{filename}}"
          },
          "org_name": {
            "type": "string",
            "description": "The name of the Github organization to use for webhooks."
          },
          "webhooks_enabled": {
            "type": "boolean",
            "description": "Whether to enable webhooks for this resource.",
            "default": false,
            "x-workforce-advanced": true
          }
        },
        "required": [
          "repo",
          "owner",
          "org_name"
        ]
      },
      "raw_text_resource_variables": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "description": "The text to use as the resource.",
            "x-workforce-multiline": true
          }
        },
        "required": [
          "text"
        ]
      },
      "simple_task_variables": {
        "type": "object",
        "properties": {
          "purpose": {
            "type": "string",
            "description": "The purpose of the task",
            "x-workforce-multiline": true
          },
          "prompt_template": {
            "type": "string",
            "description": "The template for the prompt",
            "x-workforce-multiline": true
          },
          "system_message_template": {
            "type": "string",
            "description": "Template for a guidance message to steer the model.",
            "x-workforce-multiline": true
          }
        },
        "required": [
          "prompt_template"
        ]
      },
      "structured_task_variables": {
        "type": "object",
        "properties": {
          "acceptance_critera": {
            "type": "string",
            "description": "The acceptance criteria for the task.",
            "x-workforce-multiline": true
          },
          "instructions": {
            "type": "string",
            "description": "Instructions explaining to the worker how to complete the task.",
            "x-workforce-multiline": true
          },
          "context": {
            "type": "string",
            "description": "Context that the worker should be aware of when completing the task.",
            "x-workforce-multiline": true
          },
          "definition_of_done": {
            "type": "string",
            "description": "A set of conditions that must be met for the task to be considered done.",
            "x-workforce-multiline": true
          }
        },
        "required": [
          "acceptance_critera",
          "instructions",
          "context"
        ]
      },
      "coding_tool_variables": {
        "type": "object",
        "properties": {
          "index_repo_location": {
            "type": "string",
            "description": "A URL to a mono repository without the scheme. Example github.com/my/repo."
          },
          "index_repo_branch": {
            "type": "string",
            "description": "The branch of the mono repository to use.",
            "default": "main"
          },
          "mode": {
            "type": "string",
            "description": "The mode to run the tool in.",
            "default": "local",
            "enum": [
              "local",
              "remote"
            ]
          },
          "server_url": {
            "type": "string",
            "description": "The URL of the local server to run the tool on. Required if mode is local. ",
            "default": "http://localhost:8085/vscode-extension-server"
          },
          "read_only": {
            "type": "boolean",
            "description": "Whether the tool is read only.",
            "default": false
          }
        },
        "required": [
          "index_repo_location",
          "mode"
        ]
      },
      "excalidraw_tool_variables": {
        "type": "object",
        "properties": {},
        "required": []
      },
      "github_board_ticket_tool_variables": {
        "type": "object",
        "properties": {
          "purpose": {
            "type": "string",
            "description": "The purpose of the tool",
            "x-workforce-multiline": true
          },
          "org_name": {
            "type": "string",
            "description": "The name of the Github Organization"
          },
          "project_name": {
            "type": "string",
            "description": "The name of the Github Project"
          },
          "column_name": {
            "type": "string",
            "description": "The name of the column where tickets should be placed"
          }
        },
        "required": [
          "purpose",
          "org_name",
          "project_name",
          "column_name"
        ]
      },
...
    }
  }
}